+++ packages/types/src/documents/app/table/constants.ts
@@ -1,0 +1,9 @@
+export enum RelationshipTypes {
+  ONE_TO_MANY = "one-to-many",
+  MANY_TO_ONE = "many-to-one",
+  MANY_TO_MANY = "many-to-many",
+}
+
+export enum AutoReason {
+  FOREIGN_KEY = "foreign_key",
+}
