+++ packages/types/src/documents/app/table/index.ts
@@ -1,0 +1,3 @@
+export * from "./table"
+export * from "./schema"
+export * from "./constants"
